chore: rename three-copy to four
