--- red-badger/three-copy/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "three-copy"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-thiserror = "2.0.12"
